refactor(ui): R-14 consolidate UI kits — remove components/ui/native/, single macos/ kit
UX audit finding P-009 (High): the project had two parallel UI kits:
  - components/ui/macos/ (39 components, primary, macOS-like tokens)
  - components/ui/native/ (14 components, partial duplicate, broken
    hooks.js with syntax errors)

32 source files imported from ui/native/ (mostly Card, Button, Badge,
Skeleton + animation hooks useFade/useSlide/useScale/useBreakpoint/
useTouchDevice). The native/hooks.js file had syntax errors
('const atches, setMatches]' — missing '[') which meant any import of
the hooks would crash the build; it only worked because vite tree-shaking
excluded hooks.js when only Card/Button/Badge/Skeleton were imported.

Consolidation:
1. Created frontend/src/hooks/uiAnimations.js — canonical home for the
   animation hooks (useBreakpoint, useTouchDevice, useFade, useSlide,
   useScale, useFadeIn), with the syntax errors fixed. This file is the
   single source of truth for UI animation hooks.
2. Created frontend/src/components/ui/macos/AnimatedTransition.jsx —
   moved AnimatedTransition from native/ to macos/ as a lightweight
   CSS-transition wrapper (no framer-motion dependency). Used by
   DoctorPanel.jsx (18 call sites).
3. Migrated all 32 source files:
   - 'from ../ui/native' -> 'from ../ui/macos' for Card/Button/Badge/
     Skeleton (and other components that exist in macos/)
   - 'from ../ui/native' -> 'from ../../hooks/uiAnimations' for
     useFade/useSlide/useScale/useBreakpoint/useTouchDevice
4. Deleted components/ui/native/ entirely (14 component files +
   hooks.js + index.js).
5. Deleted frontend/src/scripts/migrate-to-design-system.js,
   migrate-design-system.js, fix-import-paths.js — these were one-off
   migration helpers that referenced ui/native and are no longer
   relevant.

After this PR there is exactly ONE UI kit (components/ui/macos/) and
ONE hooks file for UI animations (hooks/uiAnimations.js). No more
'which kit should I use?' confusion, no more broken hooks.js, no more
visual inconsistency between components using different Button/Card
implementations.
