+++ frontend/src/components/admin/DesignValidator.jsx
@@ -1,5 +1,5 @@
 import { useId, useState } from 'react';
-import { Card, Button, Badge } from '../ui/native';
+import { Card, Button, Badge } from '../ui/macos';
 import {
   Select,
 } from '../ui/macos';
