+++ frontend/src/components/analytics/KPIMetrics.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { Card, Button } from '../ui/native';
+import { Card, Button } from '../ui/macos';
 import {
   TrendingUp,
   TrendingDown,
