+++ frontend/src/components/mobile/PWAInstallPrompt.jsx
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
 import { Download, X, Smartphone, Monitor } from 'lucide-react';
-import { Button, Card } from '../ui/native';
-
+import { Button, Card } from '../ui/macos';
 import logger from '../../utils/logger';
 /**
  * Компонент для предложения установки PWA
