+++ frontend/src/components/security/SMSEmail2FA.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { Card, Button } from '../ui/native';
+import { Card, Button } from '../ui/macos';
 import {
   Mail,
   Phone,
